+++ R/search_analytics.R
@@ -6,23 +6,22 @@
 #' country names.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/406-list-of-available-v4-databases-serpstatdatabaseproceduregetdatabaseinfo/}{here}.
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/ba97ni814ao9p-search-engine-short-names}{here}.
 #'
 #' @section API credits consumption: 0
 #'
 #' @param api_token (required) Serpstat API token from
-#'   \href{https://serpstat.com/users/profile/}{your profile}.
+#'   \href{https://serpstat.com/users/profile/}{your profile}. Default is Sys.getenv('SERPSTAT_API_TOKEN').
 #' @param return_method (optional) Accepted values are 'list' (default) to
 #'   return data object as list or 'df' to return data object as data.frame.
 #' @return Returns country name, se parameter value and local search engine
 #'   domain name for each country.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
-#' sst_sa_database_info(api_token)$data
+#' sst_sa_database_info()$data
 #' }
 #' @export
-sst_sa_database_info <- function(api_token, return_method = 'list'){
+sst_sa_database_info <- function(api_token = Sys.getenv('SERPSTAT_API_TOKEN'), return_method = 'list'){
   response_content <- sst_call_api_method(
     api_token  = api_token,
     api_method = 'SerpstatDatabaseProcedure.getDatabaseInfo'
@@ -34,7 +33,7 @@ sst_sa_database_info <- function(api_token, return_method = 'list'){
 #'
 #' With most API request you spend some amount of API rows. The total amount of
 #' API rows available for you is based on your
-#' \href{https://serpstat.com/pay/}{plan}. Use this method to control the amount
+#' \href{https://serpstat.com/page/pricing-plans/}{plan}. Use this method to control the amount
 #' of API rows left.
 #'
 #' @section API credits consumption: 0
@@ -45,11 +44,10 @@ sst_sa_database_info <- function(api_token, return_method = 'list'){
 #'   \href{https://serpstat.com/extension/}{Serpstat plugin} limits.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
-#' sst_sa_stats(api_token)$summary_info$left_lines
+#' sst_sa_stats()$summary_info$left_lines
 #' }
 #' @export
-sst_sa_stats <- function(api_token){
+sst_sa_stats <- function(api_token = Sys.getenv('SERPSTAT_API_TOKEN')){
   response_content <- sst_call_api_method(
     api_token  = api_token,
     api_method = 'SerpstatLimitsProcedure.getStats'
@@ -63,7 +61,7 @@ sst_sa_stats <- function(api_token){
 #' PPC, online visibility and other metrics.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/412-summarnij-otchet-po-domenu-v4-serpstatdomainproceduregetdomainsinfo/}{here}.
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/0c6vx6hzygijk-get-domains-info}{here}.
 #'
 #' @section API credits consumption: 1 per domain in request.
 #'
@@ -77,9 +75,7 @@ sst_sa_stats <- function(api_token){
 #' @return Returns aggregated stats for each domain.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
 #' sst_sa_domains_info(
-#'   api_token     = api_token,
 #'   domains       = c('amazon.com', 'ebay.com'),
 #'   se            = 'g_us',
 #'   return_method = 'df'
@@ -87,11 +83,11 @@ sst_sa_stats <- function(api_token){
 #' }
 #' @export
 sst_sa_domains_info <- function(
-    api_token,
     domains,
     se,
     sort          = NULL,
-    return_method = 'list'
+    return_method = 'list',
+    api_token = Sys.getenv('SERPSTAT_API_TOKEN')
     ){
   api_params <- list(
     domains = as.list(domains),
@@ -112,7 +108,7 @@ sst_sa_domains_info <- function(
 #' with a number of metrics for each keyword.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/584-top-search-engine-keywords-by-v4-domain-serpstatdomainproceduregetdomainkeywords/}{here}.
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/j5jbjt0rk7oei-get-domain-keywords}{here}.
 #'
 #' @section API credits consumption: 1 per keyword in response.
 #'
@@ -150,9 +146,7 @@ sst_sa_domains_info <- function(
 #' @return Returns a number of metrics for each keyword.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
 #' sst_sa_domain_keywords(
-#'   api_token     = api_token,
 #'   domain        = 'serpstat.com',
 #'   se            = 'g_us',
 #'   sort          = list(keyword_length = 'desc'),
@@ -168,7 +162,6 @@ sst_sa_domains_info <- function(
 #' }
 #' @export
 sst_sa_domain_keywords <- function(
-    api_token,
     domain,
     se,
     url           = NULL,
@@ -178,7 +171,8 @@ sst_sa_domain_keywords <- function(
     filters       = NULL,
     page          = 1,
     size          = 100,
-    return_method = 'list'
+    return_method = 'list',
+    api_token     = Sys.getenv('SERPSTAT_API_TOKEN')
     ){
   api_params <- list(
     domain        = domain,
@@ -204,7 +198,7 @@ sst_sa_domain_keywords <- function(
 #' Returns historical metrics for the domain with about two weeks between measurements.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/420-istoriya-po-domenu-v4-serpstatdomainproceduregetdomainshistory/}{here}.
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/jtjb01xqgwv8p-get-domains-history}{here}.
 #'
 #' @section API credits consumption: 1 per date in the response.
 #'
@@ -225,9 +219,7 @@ sst_sa_domain_keywords <- function(
 #' @return Returns a number of metrics for each date for the domain.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
 #' sst_sa_domain_history(
-#'   api_token     = api_token,
 #'   domain        = 'serpstat.com',
 #'   se            = 'g_us',
 #'   sort          = list(date = 'desc'),
@@ -239,7 +231,6 @@ sst_sa_domain_keywords <- function(
 #' }
 #' @export
 sst_sa_domain_history <- function(
-    api_token,
     domain,
     se,
     sort            = NULL,
@@ -247,7 +238,8 @@ sst_sa_domain_history <- function(
     page            = 1,
     size            = 100,
     during_all_time = TRUE,
-    return_method   = 'list'
+    return_method   = 'list',
+    api_token       = Sys.getenv('SERPSTAT_API_TOKEN')
     ){
   api_params <- list(
     domain          = domain,
@@ -272,7 +264,7 @@ sst_sa_domain_history <- function(
 #'   number of keywords, and Facebook shares.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/588-domain-top-urls-v4-serpstatdomainproceduregettopurls/}{here}.
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/94ol4y7b9zje4-get-top-urls}{here}.
 #'
 #' @section API credits consumption: 1 per returned page.
 #'
@@ -291,9 +283,7 @@ sst_sa_domain_history <- function(
 #' @return Returns domain top pages with their metrics.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
 #' sst_sa_domain_top_pages(
-#'   api_token     = api_token,
 #'   domain        = 'serpstat.com',
 #'   se            = 'g_us',
 #'   sort          = list(organic_keywords = 'desc'),
@@ -305,14 +295,14 @@ sst_sa_domain_history <- function(
 #' }
 #' @export
 sst_sa_domain_top_pages <- function(
-    api_token,
     domain,
     se,
     sort          = NULL,
     filters       = NULL,
     page          = 1,
     size          = 100,
-    return_method = 'list'
+    return_method = 'list',
+    api_token     = Sys.getenv('SERPSTAT_API_TOKEN')
     ){
   api_params <- list(
     domain  = domain,
@@ -335,7 +325,7 @@ sst_sa_domain_top_pages <- function(
 #' Returns organic competitors for the domain with their key metrics.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/588-domain-top-urls-v4-serpstatdomainproceduregettopurls/}{here}.
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/qzd94b49vdkyb-get-organic-competitors-page}{here}.
 #'
 #' @section API credits consumption: 1 per returned page.
 #'
@@ -352,9 +342,7 @@ sst_sa_domain_top_pages <- function(
 #' @return Returns a number of metrics for each organic competitor.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
 #' sst_sa_domain_organic_competitors(
-#'   api_token     = api_token,
 #'   domain        = 'serpstat.com',
 #'   se            = 'g_us',
 #'   sort          = list(relevance = 'desc'),
@@ -365,13 +353,13 @@ sst_sa_domain_top_pages <- function(
 #' }
 #' @export
 sst_sa_domain_organic_competitors <- function(
-    api_token,
     domain,
     se,
     sort          = NULL,
     page          = 1,
     size          = 100,
-    return_method = 'list'
+    return_method = 'list',
+    api_token     = Sys.getenv('SERPSTAT_API_TOKEN')
     ){
   api_params <- list(
     domain = domain,
@@ -388,13 +376,58 @@ sst_sa_domain_organic_competitors <- function(
   sst_return_check(response_content, return_method)
 }
 
+#' Number of keywords for each region for a domain
+#'
+#' Returns the number of keywords for each region for a domain. Regions with no keywords
+#' are not returned.
+#'
+#' @section API docs:
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/t84jvml38pg1q-get-regions-count}{here}.
+#'
+#' @section API credits consumption: 1 per returned region.
+#'
+#' @param domain (required) A domain name to analyze.
+#' @param sort (optional) A field to sort the response. Allowed values: keywords_count, db_name, country_name_en, google_domain. Default: keywords_count.
+#' @param order (optional) The order of sorting. Allowed values: asc, desc. Default: desc.
+#' @inheritParams sst_sa_database_info
+#' @return Returns a data frame with number of keywords for each region for a domain.
+#' @examples
+#' \dontrun{
+#' sst_sa_domain_keywords_per_region(
+#'   domain        = 'serpstat.com',
+#'   sort          = 'keywords_count',
+#'   order         = 'desc',
+#'   return_method = 'df'
+#' )$data
+#' }
+#' @export
+sst_sa_domain_keywords_per_region <- function(
+    domain,
+    sort          = 'keywords_count',
+    order         = 'desc',
+    return_method = 'list',
+    api_token     = Sys.getenv('SERPSTAT_API_TOKEN')
+){
+  api_params <- list(
+    domain = domain,
+    sort   = sort,
+    order  = order
+  )
+  response_content <- sst_call_api_method(
+    api_token  = api_token,
+    api_method = 'SerpstatDomainProcedure.getRegionsCount',
+    api_params = api_params
+  )
+  sst_return_check(response_content, return_method)
+}
+
 #' Keywords summary
 #'
 #' Returns a number of metrics for each keyword like search volume, CPC and
 #' competition level.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/600-keyword-info-v4-serpstatkeywordproceduregetkeywordsinfo/}{here}.
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/2brcbdgmr6rcx-get-keywords-info}{here}.
 #'
 #' @section API credits consumption: 1 per keyword in request.
 #'
@@ -408,9 +441,7 @@ sst_sa_domain_organic_competitors <- function(
 #' @return Returns a number of metrics for each keyword.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
 #' sst_sa_keywords_info(
-#'   api_token     = api_token,
 #'   keywords      = c('seo', 'ppc', 'serpstat'),
 #'   se            = 'g_us',
 #'   sort          = list(cost = 'asc'),
@@ -419,11 +450,11 @@ sst_sa_domain_organic_competitors <- function(
 #' }
 #' @export
 sst_sa_keywords_info <- function(
-    api_token,
     keywords,
     se,
     sort          = NULL,
-    return_method = 'list'
+    return_method = 'list',
+    api_token     = Sys.getenv('SERPSTAT_API_TOKEN')
     ){
   api_params <- list(
     keywords = as.list(keywords),
@@ -445,7 +476,7 @@ sst_sa_keywords_info <- function(
 #' level.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/592-keywords-for-v4-words-serpstatkeywordproceduregetkeywords/}{here}.
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/w7jh5sk9kc0cm-get-keywords}{here}.
 #'
 #' @section API credits consumption: 1 per keyword in response.
 #'
@@ -466,9 +497,7 @@ sst_sa_keywords_info <- function(
 #' @return Returns a number of metrics for each keyword.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
 #' sst_sa_keywords(
-#'   api_token     = api_token,
 #'   keyword       = 'serpstat',
 #'   se            = 'g_us',
 #'   minusKeywords = c('free'),
@@ -480,7 +509,6 @@ sst_sa_keywords_info <- function(
 #' }
 #' @export
 sst_sa_keywords <- function(
-    api_token,
     keyword,
     se,
     minusKeywords = NULL,
@@ -488,7 +516,8 @@ sst_sa_keywords <- function(
     filters       = NULL,
     page          = 1,
     size          = 100,
-    return_method = 'list'
+    return_method = 'list',
+    api_token     = Sys.getenv('SERPSTAT_API_TOKEN')
     ){
   api_params <- list(
     keyword       = keyword,
@@ -512,37 +541,97 @@ sst_sa_keywords <- function(
 #' organic results, paid results and different types of SERP features.
 #'
 #' @section API docs:
-#'  Check all the values for request and response fields \href{https://serpstat.com/api/598-keyword-top-v4-serpstatkeywordproceduregetkeywordtop/}{here}.
-#'
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/env4osjb1ev9c-get-keyword-top}{here}.
 #'
 #' @section API credits consumption: 1 per URL in response.
 #'
 #' @param keyword (required) A keyword to search for.
 #' @param se (required) Search engine alias (db_name) returned by
 #'     \code{\link{sst_sa_database_info}}.
-#' @param  top_size (optional) Set the number of URLs to get in response.
+#' @param sort (optional) A field to sort the response. See API docs for more details.
+#' @param size (optional) Response page size. Should be => 10 and <=100. Default is 100.
 #' @inheritParams sst_sa_database_info
 #' @return Returns a list with the data about search engine results page for the keyword.
 #' @examples
 #' \dontrun{
-#' api_token <- Sys.getenv('SERPSTAT_API_TOKEN')
 #' sst_sa_keyword_top(
-#'   api_token = api_token,
 #'   keyword   = 'serpstat',
 #'   se        = 'g_us',
-#'   top_size  = 10
+#'   size  = 10
 #' )
 #' }
 #' @export
-sst_sa_keyword_top <- function(api_token, keyword, se, top_size = 100) {
+sst_sa_keyword_top <- function(
+    keyword,
+    se,
+    sort      = NULL,
+    size      = 100,
+    api_token = Sys.getenv('SERPSTAT_API_TOKEN')
+    ){
   response_content <- sst_call_api_method(
     api_token  = api_token,
-    api_method = 'SerpstatKeywordProcedure.getKeywordTop',
+    api_method = 'SerpstatKeywordProcedure.getKeywordFullTop',
     api_params = list(
       keyword = keyword,
       se      = se,
-      filters = list(top_size = top_size)
+      sort    = sort,
+      size    = size
     )
   )
   sst_return_check(response_content, return_method = 'list')
+}
+
+#' Related keywords for a keyword
+#'
+#' Returns a list of semantically related keywords for a given keyword.
+#'
+#' @section API docs:
+#'  Check all the values for request and response fields \href{https://api-docs.serpstat.com/docs/serpstat-public-api/7yidbv07n5q9u-get-related-keywords}{here}.
+#'
+#' @section API credits consumption: 1 per keyword in response.
+#'
+#' @param keyword (required) A keyword to search for.
+#' @param se (required) Search engine alias (db_name) returned by \code{\link{sst_sa_database_info}}.
+#' @param with_intents (optional) TRUE if keyword intent should be retrieved. This parameter works for g_ua and g_us database only.
+#' @param filters (optional) A list of filtering options. See API docs for more details.
+#' @param sort (optional) A field to sort the response. See API docs for more details.
+#' @param page (optional) Response page number if there are many pages in response.
+#' @param size (optional) Response page size.
+#' @inheritParams sst_sa_database_info
+#' @return Returns related keywords for a keyword and their metrics.
+#' @examples
+#' \dontrun{
+#' sst_sa_related_keywords(
+#'   keyword       = 'serpstat',
+#'   se            = 'g_us',
+#'   return_method = 'df'
+#' )$data
+#' }
+#' @export
+sst_sa_related_keywords <- function(
+    keyword,
+    se,
+    with_intents  = FALSE,
+    filters       = NULL,
+    sort          = NULL,
+    page          = 1,
+    size          = 100,
+    return_method = 'list',
+    api_token     = Sys.getenv('SERPSTAT_API_TOKEN')
+    ){
+  api_params <- list(
+    keyword       = keyword,
+    se            = se,
+    withIntents   = with_intents,
+    filters       = filters,
+    sort          = sort,
+    page          = page,
+    size          = size
+  )
+  response_content <- sst_call_api_method(
+    api_token  = api_token,
+    api_method = 'SerpstatKeywordProcedure.getRelatedKeywords',
+    api_params = api_params
+    )
+  sst_return_check(response_content, return_method)
 }